+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 3.25.0 
+* Add `evidenceSubmittable` to `Dispute`
+* Add `merchantTokenIdentifier`, `sourceCardLast4` to `ApplePayCard` and `ApplePayDetails`
+* Add `threeDSecureAuthenticationId` to `CustomerRequest` and `PaymentMethodRequest`
+* Add `processingOverrides` to `TransactionOptionsRequest`
+* Add the following fields to `processingOverrides`
+  * customerEmail
+  * customerFirstName
+  * customerLastName
+  * customerTaxIdentifier
+
 ## 3.24.0
 * Allow vaulting of Android Pay network tokens via Customer.create/update
   * Add `AndroidPayNetworkTokenRequest`